pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ioc-enricher"
version = "0.1.0"
description = "CLI tool to enrich threat intelligence IOCs via VirusTotal, AbuseIPDB, and Shodan"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.scripts]
ioc-enricher = "ioc_enricher.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ioc_enricher*"]
